chore: refactor browser ui logic
